+++ test/docfx.Test/lib/FileGlobTest.cs
@@ -4,8 +4,7 @@
 using System;
 using System.IO;
 using System.Linq;
-using Newtonsoft.Json;
-using Newtonsoft.Json.Linq;
+
 using Xunit;
 
 namespace Microsoft.Docs.Build
